% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnorm_mean_sd.R
\name{dlnorm_mean_sd}
\alias{dlnorm_mean_sd}
\title{This function provides the density function of the lognormal distribution based on mean and stddev}
\usage{
dlnorm_mean_sd(x, mean, sd, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{mean}{vector of means}

\item{sd}{vector of standard deviations}
}
\description{
This function provides the density function of the lognormal distribution based on mean and stddev
}
