% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_mean_sd.R
\name{pgamma_mean_sd}
\alias{pgamma_mean_sd}
\title{This function provides the distribution function of the gamma distribution based on mean and stddev}
\usage{
pgamma_mean_sd(x, mean, sd, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{mean}{vector of means}

\item{sd}{vector of standard deviations}
}
\description{
This function provides the distribution function of the gamma distribution based on mean and stddev
}
