% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panjer_distributions.R
\name{ppanjer}
\alias{ppanjer}
\title{This function provides the distribution function of the Panjer distribution class based on mean and dispersion/overdispersion/contagion}
\usage{
ppanjer(x, mean, dispersion, overdispersion = dispersion - 1,
  contagion = overdispersion/mean, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{mean}{vector of means}

\item{dispersion}{vector of dispersions}

\item{overdispersion}{vector of overdispersions}

\item{contagion}{vector of contagions}
}
\description{
This function provides the distribution function of the Panjer distribution class based on mean and dispersion/overdispersion/contagion
}
