% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnorm_mean_sd.R
\name{qlnorm_mean_sd}
\alias{qlnorm_mean_sd}
\title{This function provides the quantile function of the lognormal distribution based on mean and stddev}
\usage{
qlnorm_mean_sd(p, mean, sd, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{vector of probabilities}

\item{mean}{vector of means}

\item{sd}{vector of standard deviations}
}
\description{
This function provides the quantile function of the lognormal distribution based on mean and stddev
}
