% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_mean_sd.R
\name{rgamma_mean_sd}
\alias{rgamma_mean_sd}
\title{This function provides random generation for the gamma distribution based on mean and stddev}
\usage{
rgamma_mean_sd(n, mean, sd, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{n}{number of observations}

\item{mean}{vector of means}

\item{sd}{vector of standard deviations}
}
\description{
This function provides random generation for the gamma distribution based on mean and stddev
}
