% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnorm_mean_sd.R
\name{rlnorm_mean_sd}
\alias{rlnorm_mean_sd}
\title{This function provides random generation for the lognormal distribution based on mean and stddev}
\usage{
rlnorm_mean_sd(n, mean, sd, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{n}{number of observations}

\item{mean}{vector of means}

\item{sd}{vector of standard deviations}
}
\description{
This function provides random generation for the lognormal distribution based on mean and stddev
}
