% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panjer_distributions.R
\name{rpanjer}
\alias{rpanjer}
\title{This function provides random generation the Panjer distribution class based on mean and dispersion/overdispersion/contagion}
\usage{
rpanjer(n, mean, dispersion, overdispersion = dispersion - 1,
  contagion = overdispersion/mean, log = FALSE)
}
\arguments{
\item{n}{number of observations}

\item{mean}{vector of means}

\item{dispersion}{vector of dispersions}

\item{overdispersion}{vector of overdispersions}

\item{contagion}{vector of contagions}
}
\description{
This function provides random generation the Panjer distribution class based on mean and dispersion/overdispersion/contagion
}
